# Optimize the performance of dot by using universal intrinsics in X86/ARM