## Introduction

Compare with develop branch, when operating float dot, the new implementation has about 3x faster in AVX2, about 10%~30% faster in SSE2/SSE3/NEON. 

## Benchmark

  Each test has run for 10 loops and then take the average result, The radio is the MFlops division of Baseline and optimized branch(*The bigger* the radio is, the *better* performance is achieved).

- X86 BaseLine(auto vectorized)

| data size(10^3) | develop(float)               |
| --------------- | ---------------------------- |
| 4000            | 772.81 MFlops   0.010352 sec |
| 8000            | 769.16 MFlops   0.020802 sec |

- X86-SSE2 enabled

  | data size(10^3) | usimd-floatdot(float)        | ratio |
  | --------------- | ---------------------------- | ----- |
  | 4000            | 819.17 MFlops   0.009766 sec | 1.05  |
  | 8000            | 815.09 MFlops   0.019630 sec | 1.05  |
  
- X86-SSE3 enabled

  | data size(10^3) | usimd-floatdot(float)        | ratio |
  | --------------- | ---------------------------- | ----- |
  | 4000            | 853.32 MFlops   0.009375 sec | 1.10  |
  | 8000            | 848.88 MFlops   0.018848 sec | 1.10  |
  
- X86-AVX2 enabled

  | data size(10^3) | usimd-floatdot(float)       | ratio |
  | --------------- | --------------------------- | ----- |
  | 4000            | 2482.31 MFlops 0.003223 sec | 3.21  |
  | 8000            | 2307.47 MFlops 0.006934 sec | 2.99  |
  
- ARM BaseLine(dot.s)

  | data size(10^3) | develop(float)                |
  | --------------- | ----------------------------- |
  | 4000            | 1009.15 MFlops   0.007927 sec |
  | 8000            | 1042.69 MFlops   0.015345 sec |
  
- ARM-Neon enabled

  | data size | usimd-floatdot(float)         | ratio |
  | --------- | ----------------------------- | ----- |
  | 4000      | 1311.65 MFlops   0.006099 sec | 1.30  |
  | 8000      | 1274.01 MFlops   0.012559 sec | 1.22  |

## System Info

|           | Arm                                                          | x86                                      |
| --------- | ------------------------------------------------------------ | ---------------------------------------- |
| Hardware  | KunPeng                                                      |                                          |
| Processor | ARMv8 2.6GMHZ 8 processors                                   | Intel(R) Xeon(R) Gold 6161 CPU @ 2.20GHz |
| OS        | Linux ecs-9d50 4.19.36-vhulk1905.1.0.h276.eulerosv2r8.aarch64 | Windows Server 2008 R2 Enterprise        |
| Compiler  | gcc (GCC) 7.3.0                                              | MSVC14.06                                |

## Co-Author

@seiko2plus 

## Reference

1. https://software.intel.com/sites/landingpage/IntrinsicsGuide/
2. https://www.thetopsites.net/article/58234904.shtml
3. https://github.com/numpy/numpy/tree/master/numpy/core/src/common/simd
4. https://github.com/opencv/opencv/tree/master/modules/core/include/opencv2/core/hal