@Qiyu8 2020-10-09
It's looks like `-msse3` is not enabled during compiling while HAVE_SSE3 is detected.