@martin-frbg 2020-10-09
Might well be - so far any uses of particular SSE levels must have been in assembly code where the HAVE_SSE3 would only be used to guard wether the mnemonic should be used or not. For using them with intrinsics it would be necessary to add the compiler option in Makefile.x86_64, kernel/Makefile and cmake/system.cmake (as it is currenlty done to add `-mavx2` for HAVE_AVX2)